+++ android/src/car/superfun/game/AndroidLauncher.java
@@ -22,6 +22,7 @@ import java.util.ArrayList;
 
 import car.superfun.game.car.OpponentCarController;
 import car.superfun.game.googleGamePlayServices.Communicator;
+import car.superfun.game.googleGamePlayServices.GoogleGameServices;
 import car.superfun.game.googleGamePlayServices.SetUpGame;
 import car.superfun.game.googleGamePlayServices.SignIn;
 import car.superfun.game.states.NewState;
