+++ android/src/car/superfun/game/actors/CheckboxActor.java
@@ -1,4 +1,4 @@
-package car.superfun.game.actor;
+package car.superfun.game.actors;
 
 /**
  * Created by Jonas on 10.04.2018.
