+++ android/src/car/superfun/game/actors/DelayedButtonActor.java
@@ -1,4 +1,4 @@
-package car.superfun.game.actor;
+package car.superfun.game.actors;
 
 import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Batch;
