+++ android/src/car/superfun/game/menus/LoadingScreen.java
@@ -2,16 +2,14 @@ package car.superfun.game.menus;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.GL20;
-import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.math.Circle;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.InputListener;
 import com.badlogic.gdx.scenes.scene2d.Stage;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 
-import car.superfun.game.actor.DelayedButtonActor;
-import car.superfun.game.actor.LoadingActor;
+import car.superfun.game.actors.DelayedButtonActor;
+import car.superfun.game.actors.LoadingActor;
 import car.superfun.game.states.GameStateManager;
 import car.superfun.game.states.State;
