+++ android/src/car/superfun/game/menus/LoginMenu.java
@@ -5,7 +5,7 @@ import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 
-import car.superfun.game.GoogleGameServices;
+import car.superfun.game.googleGamePlayServices.GoogleGameServices;
 import car.superfun.game.states.GameStateManager;
 import car.superfun.game.states.State;
