+++ android/src/car/superfun/game/menus/ScoreFormatter.java
@@ -1,0 +1,6 @@
+package car.superfun.game.menus;
+
+public interface ScoreFormatter {
+    String formatScore(int score);
+    String scoreString();
+}
