+++ android/src/car/superfun/game/menus/SettingsMenu.java
@@ -2,7 +2,6 @@ package car.superfun.game.menus;
 
 import com.badlogic.gdx.Gdx;
 import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 import com.badlogic.gdx.scenes.scene2d.InputEvent;
 import com.badlogic.gdx.scenes.scene2d.InputListener;
@@ -14,7 +13,7 @@ import com.badlogic.gdx.scenes.scene2d.ui.Table;
 import com.badlogic.gdx.utils.Align;
 import com.badlogic.gdx.utils.viewport.ScreenViewport;
 
-import car.superfun.game.actor.ButtonActor;
+import car.superfun.game.actors.ButtonActor;
 import car.superfun.game.states.GameStateManager;
 import car.superfun.game.states.State;
